+++ src/app.ts
@@ -3,6 +3,9 @@ import express from 'express'
 import bodyParser from 'body-parser'
 import cors from 'cors'
 import reports from './reports/router'
+import streams from './streams/router'
+import projects from './projects/router'
+import './events/consumer'
 import docs from './docs'
 import { jwtCheck, parseUserData } from './common/auth'
 import { errorHandler, notFoundHandler } from './common/error-handlers'
@@ -15,7 +18,8 @@ app.use(bodyParser.json())
 app.use('/docs', docs)
 app.use(jwtCheck, parseUserData)
 
-// app.use('/projects', projects)
+app.use('/streams', streams)
+app.use('/projects', projects)
 app.use('/reports', reports)
 
 app.use(errorHandler)