+++ src/common/core-api/types.ts
@@ -1,0 +1,18 @@
+export interface ForwardedResponse<T> {
+  data: T[]
+  headers: any
+}
+
+export interface ProjectResponse {
+  id: string
+  name: string
+}
+
+export interface StreamResponse {
+  id: string
+  name: string
+  latitude: number
+  longitude: number
+  altitude: number
+  project: ProjectResponse | null
+}