+++ src/streams/types.ts
@@ -1,0 +1,16 @@
+import { StreamResponse } from '../types'
+
+export interface StreamQuery {
+  projects: string[]
+  only_public: boolean
+  keyword: string
+  limit: number
+  offset: number
+  sort: string
+  fields: string[]
+  with_events_count: boolean
+}
+
+export interface StreamResponseWithEventsCount extends StreamResponse {
+  eventsCount: number
+}