+++ CHANGELOG.md
@@ -1,5 +1,8 @@
-## 0.0.1 (2021-09-10)
+## 0.0.1 (2021-09-XX)
 
 ### Features
-* Setup project structure, database with models; test environment ([CE-1359](https://jira.rfcx.org/browse/CE-1359))
-* Add endpoint for report creation with tests ([CE-1360](https://jira.rfcx.org/browse/CE-1360))
+- Setup project structure, database with models; test environment ([CE-1359](https://jira.rfcx.org/browse/CE-1359))
+- Add endpoint for report creation with tests ([CE-1360](https://jira.rfcx.org/browse/CE-1360))
+- Add SQS consumer for Event creation from Core API (including Classification) ([CE-1394](https://jira.rfcx.org/browse/CE-1394))
+- Add tests for Event creation
+- Add endpoint to get list of sites (including events count since last report), add tests ([CE-1394](https://jira.rfcx.org/browse/CE-1394))